streamlit-chatbot-app/requirements.txt
google-generativeai
streamlit
mysql-connector-python